Add status color util
